feat: update transfer (#72)
* feat: remove useless code

* feat: update discover new amount

* feat: update retry times of voting

* feat: fetch portkey token after connected

* feat: remove logs

* feat: update fetch token

* feat: update transfer
