+++ src/routes/index.tsx
@@ -6,11 +6,7 @@ import Home from "@/pageComponents/Home";
 import CreatePoll from "@/pageComponents/CreatePoll";
 import PollDetail from "@/pageComponents/PollDetail";
 import { postWithToken } from "@/hooks/useData";
-import { useRequest } from "ahooks";
 import { chainId } from "@/constants/app";
-import { nftSymbol } from "@/config";
-import { useConnectWallet } from "@aelf-web-login/wallet-adapter-react";
-import { useUserContext } from "@/provider/UserProvider";
 import { isInTelegram } from "@/utils/isInTelegram";
 
 const routes = [
@@ -35,48 +31,6 @@ const routes = [
 
 const AppRoutes: React.FC = () => {
   const lp = useLaunchParams();
-  const { isConnected, walletInfo: wallet } = useConnectWallet();
-  const { fetchTokenAndData } = useUserContext();
-
-  const fetchTransfer = async (cancel: () => void) => {
-    const { data } = await postWithToken("/api/app/token/transfer", {
-      chainId,
-      symbol: nftSymbol,
-    });
-    if (!data) {
-      if (isInTelegram()) {
-        window.location.reload();
-      } else {
-        fetchTokenAndData();
-      }
-    } else {
-      cancel();
-    }
-  };
-
-  const { run: fetchTransferStatus, cancel } = useRequest(
-    async () => {
-      try {
-        const { data } = await postWithToken("/api/app/token/transfer/status", {
-          chainId,
-          address: wallet?.address,
-          symbol: nftSymbol,
-        });
-        const { isClaimedInSystem } = data || {};
-        if (!data || !isClaimedInSystem) {
-          fetchTransfer(cancel);
-        } else {
-          cancel();
-        }
-      } catch (error) {
-        console.error(error);
-      }
-    },
-    {
-      manual: true,
-      pollingInterval: 1000,
-    }
-  );
 
   const saveUserInfo = async () => {
     try {
@@ -101,12 +55,6 @@ const AppRoutes: React.FC = () => {
     }
   }, []);
 
-  useEffect(() => {
-    if (isConnected && wallet?.address) {
-      fetchTransferStatus();
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [isConnected, wallet?.address]);
   return (
     <AppRoot
       appearance="dark"
